requirements.txt
django>=4.2
django-recaptcha
django-hcaptcha
django-turnstile
python-dotenv
